+++ CHANGELOG.md
@@ -11,6 +11,7 @@
     to `dist/parsley.remote.js` and `dist/parsley.remote.min.js`
   - changed the way `asyncValidate` handle events
   - fix warning if parsley called on a page without elements to validate (#562)
+  - fixed `ParsleyUtils.attr()` attribute checking (#564)
 
 ## 2.0.0-rc4